Reworked veneer support.
Veneers themselves are now generated by pmbfd.

Every loaded object (+ system table) calls cexpSymTblFixup()
which in-turn calls pmbfd_make_veneers() to create veneers
and attach them to the symbol's 'xtra' field.

When such a symbol is referred by a relocation then the
extra info is passed to pmbfd_perform_reloc() where it
can be used to locate a proper veneer.